поправил ошибки
